Reorg package names for ZTS change log store
