+++ servers/zts/conf/zts.properties
@@ -118,7 +118,7 @@ athenz.zts.cert_signer_factory_class=com.yahoo.athenz.zts.cert.impl.SelfCertSign
 # Specifies the factory class that implements ChangeLogStoreFactory
 # interface used by the ZTS Server to retrieve the latest changes
 # from the ZMS Server and save them locally
-#athenz.zts.change_log_store_factory_class=com.yahoo.athenz.zts.store.file.ZMSFileChangeLogStoreFactory
+#athenz.zts.change_log_store_factory_class=com.yahoo.athenz.zts.store.impl.ZMSFileChangeLogStoreFactory
 
 # Specifies the directory for storing zms domain json documents when
 # ZMSFileChangeLogStoreFactory is configured for the change log factory
