+++ servers/zts/src/main/java/com/yahoo/athenz/zts/store/impl/S3ChangeLogStore.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.yahoo.athenz.zts.store.s3;
+package com.yahoo.athenz.zts.store.impl;
 
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
