+++ servers/zts/src/test/java/com/yahoo/athenz/zts/store/impl/MockZMSFileChangeLogStoreFactory.java
@@ -13,23 +13,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.yahoo.athenz.zts.store.file;
+package com.yahoo.athenz.zts.store.impl;
 
-import java.io.File;
 import java.security.PrivateKey;
 
 import com.yahoo.athenz.zts.store.ChangeLogStore;
 import com.yahoo.athenz.zts.store.ChangeLogStoreFactory;
 import com.yahoo.athenz.zts.store.CloudStore;
 
-public class ZMSFileChangeLogStoreFactory implements ChangeLogStoreFactory {
+public class MockZMSFileChangeLogStoreFactory implements ChangeLogStoreFactory {
 
-    private static final String ZTS_DATA_STORE = "zts_store";
-
+    @Override
     public ChangeLogStore create(String ztsHomeDir, PrivateKey privateKey,
             String privateKeyId, CloudStore cloudStore) {
-        
-        return new ZMSFileChangeLogStore(ztsHomeDir + File.separator + ZTS_DATA_STORE, privateKey,
-                privateKeyId);
+        return new MockZMSFileChangeLogStore(ztsHomeDir, privateKey, privateKeyId);
     }
 }
